Rename JsonApiResource interface as IncludesRelatedResources
